refactor: server.js
