+++ BackEnd/middlewares/errorHandler.js
@@ -1,0 +1,4 @@
+module.exports = (err, req, res, next) => {
+  console.log(err);
+  return res.status(500).json({ success: false, error: err });
+};
