+++ BackEnd/middlewares/requestLogger.js
@@ -1,0 +1,4 @@
+module.exports = (req, res, next) => {
+  console.log(req.ip, "- ", req.url);
+  next();
+};
