+++ BackEnd/middlewares/urlLogger.js
@@ -1,0 +1,4 @@
+module.exports = (req, res, next) => {
+  console.log(req.url);
+  next();
+};
